errors: move ConnectionError to scylla crate
Again, it was defined in scylla-cql crate because
ConnectionPoolError directly depends on it.
